Adjust #current_user and remove #logged_in? method
Used session[:user_id] instead of #logged_in? to find instance of user.
The conditional statement for displaying the navbar kept
calling #logged_in?, which redirected to root_path, which called
would call #logged_in?. If user wasn't logged in, it caused a loop.
Once the was resolved, I was able to remove the #logged_in? method
because it was not needed.